完成初始資料切換
