+++ config/init/b2b/index.js
@@ -5,13 +5,15 @@ import order from './order/index.js';
 import orderProduct from './orderproduct/index.js';
 import orderPayment from './orderpayment/index.js';
 import orderStatus from './orderstatus/index.js';
+import menuItem from './orderstatus';
 
 module.exports.init = async function(){
+  console.log("=== init b2b data ===");
   await product.init();
   await supplier.init();
   await orderStatus.init();
   await order.init();
   await orderProduct.init();
   await supplierShipOrder.init();
-  await orderPayment.init();
+  await menuItem.init();
 }
