--- config/init/default/index.js
@@ -1,7 +1,0 @@
-module.exports.init = async () => {
-  try {
-
-  } catch (e) {
-    console.error(e);
-  }
-};
