+++ config/init/labfnp/allpay/index.js
