+++ config/init/labfnp/facebook/index.js
