+++ config/init/labfnp/message/index.js
