+++ config/init/labfnp/other/data/ScentDetail.json
